Add session checking

+++ src/main/java/pl/agh/edu/erasmus_system/controller/contract_coordinators/ContractCoordinatorController.java
@@ -15,6 +15,7 @@ import pl.agh.edu.erasmus_system.model.Session;
 import pl.agh.edu.erasmus_system.repository.ContractCoordinatorRepository;
 import pl.agh.edu.erasmus_system.repository.ContractRepository;
 import pl.agh.edu.erasmus_system.repository.SessionRepository;
+import pl.agh.edu.erasmus_system.service.SessionService;
 
 import java.util.*;
 
@@ -32,13 +33,16 @@ public class ContractCoordinatorController {
     @Autowired
     private SessionRepository sessionRepository;
 
+    @Autowired
+    private SessionService sessionService;
+
     /**
      * @param requestBody- body of request with params: email and password
      * @return OK (200) when credentials are correct
      * UNAUTHORIZED (401) when credentials are wrong
      */
     @RequestMapping(value ="/login", method = RequestMethod.POST)
-    public ResponseEntity<String> login(@RequestBody LoginRequestBody requestBody) {
+    public ResponseEntity login(@RequestBody LoginRequestBody requestBody) {
         String email = requestBody.getEmail();
         String password = requestBody.getPassword();
 
@@ -48,14 +52,19 @@ public class ContractCoordinatorController {
             if(PasswordManagement.check(password, coordinator.getHash())){
                 Session newSession = SessionGenerator.getNewSession(coordinator, sessionRepository);
                 sessionRepository.save(newSession);
-                return new ResponseEntity<>("\"" + newSession.getCode() + "\"", HttpStatus.OK);
+                return new ResponseEntity(HttpStatus.OK);
             }
         }
-        return new ResponseEntity<>("", HttpStatus.UNAUTHORIZED);
+        return new ResponseEntity(HttpStatus.UNAUTHORIZED);
     }
 
     @RequestMapping(value = "/delete/all", method = RequestMethod.GET)
     public ResponseEntity deleteAllContractCoordinators() {
+        ContractsCoordinator coordinator = sessionService.getCoordinatorOf("sessionCode");
+        if (coordinator == null) {
+            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+        }
+
         List<ContractsCoordinator> contractsCoordinators = contractCoordinatorRepository.findAll();
         for (var contractCoordinator : contractsCoordinators) {
             contractCoordinatorRepository.delete(contractCoordinator);
@@ -86,6 +95,11 @@ public class ContractCoordinatorController {
     }
 
     private ResponseEntity<ContractCoordinatorResponseBody> getContractCoordinatorResponseEntity(List<ContractsCoordinator> contractsCoordinators) {
+        ContractsCoordinator coordinator = sessionService.getCoordinatorOf("sessionCode");
+        if (coordinator == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         ContractCoordinatorResponseBody response = new ContractCoordinatorResponseBody();
 
         for (ContractsCoordinator contractsCoordinator : contractsCoordinators) {
@@ -101,6 +115,4 @@ public class ContractCoordinatorController {
         }
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
-
-
 }
